refactor(editor): split CodeEditorView.swift into focused files
Breaks up the 2186-LOC monolith into five files along natural seams:

- GutterTextView.swift (~540) — NSTextView subclass: gutter inset,
  current-line highlight, inline diff + blame drawing, auto-indent,
  comment toggling, Escape-to-collapse.
- EditorContainerViews.swift (~120) — EditorScrollView (find-bar
  offset tracking), EditorContainerView (scroll / gutter / minimap
  layout), GoToRequest.
- CodeEditorView+Coordinator.swift (~1090) — the nested Coordinator
  (NSTextViewDelegate / NSTextStorageDelegate / NSLayoutManagerDelegate),
  moved into an 'extension CodeEditorView' so callers still refer to
  it as CodeEditorView.Coordinator.
- NSImage+Tinted.swift (~25) — the shared tint helper used by the
  gutter blame icon.
- CodeEditorView.swift (~510) — now only the SwiftUI wrapper:
  makeNSView / updateNSView / makeCoordinator, static initial-range
  estimation helpers, and applyViewportHighlighting.

Strictly mechanical: no behavior change, no public API change. The
only access-level adjustment is promoting editorFont and
applyViewportHighlighting from private to internal so the Coordinator,
now in a separate file, can still reach them.

Fixes #755
